+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 -------------------
 
 - added `@abc.abstractmethod` decorator where appropriate
+- added `stopped` flag to `Session` to indicate that the execution is over,
+  which is monitored by the `seppl.io.execute(...)` function
 
 
 0.2.21 (2025-08-19)